@@ -1,6 +1,6 @@
 {
 	"name": "diceslice",
-	"version": "0.0.1",
+	"version": "0.1.0",
 	"private": true,
 	"type": "module",
 	"scripts": {
@@ -17,29 +17,29 @@
 	},
 	"devDependencies": {
 		"@playwright/test": "^1.46.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@tailwindcss/forms": "^0.5.7",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/forms": "^0.5.10",
+		"@tailwindcss/vite": "^4.0.9",
 		"@tsconfig/svelte": "^5.0.4",
 		"@types/eslint": "^8.56.0",
 		"@types/node": "^20.11.24",
 		"@typescript-eslint/eslint-plugin": "^7.0.0",
 		"@typescript-eslint/parser": "^7.0.0",
-		"@vitest/browser": "^1.3.1",
+		"@vitest/browser": "^3.0.7",
 		"autoprefixer": "^10.4.20",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^3.0.2",
 		"playwright": "^1.42.1",
-		"postcss": "^8.4.32",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.4.17",
+		"prettier-plugin-svelte": "^3.2.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^3.7.0",
+		"tailwindcss": "^4.0.9",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
-		"vitest": "^1.2.0"
+		"vite": "^6.2.0",
+		"vitest": "^3.0.7"
 	},
 	"dependencies": {
 		"comlink": "^4.4.1",
