@@ -1,7 +1,8 @@
+import { mount } from 'svelte'
 import App from './App.svelte'
 
-const app = new App({
+const root = mount(App, {
   target: document.getElementById('app')!
 })
 
-export default app 
+export default root 
