@@ -1,18 +1,24 @@
+/// <reference types="@vitest/browser/providers/playwright" />
+
 import { defineConfig } from 'vitest/config'
 import { resolve } from 'path'
 
 export default defineConfig({
     test: {
         browser: {
-            enabled: true,
             headless: true,
-            name: 'chromium',
-            provider: 'playwright'
+            provider: 'playwright',
+            enabled: true,
+            instances: [
+                {
+                    browser: 'chromium',
+                }
+            ]
         },
         include: ['tests/unit/lib/**/*.{test,spec}.{js,ts}'],
         coverage: {
             provider: 'istanbul'
-        }
+        },
     },
     resolve: {
         alias: {
